Miner 2149 Vr. 3.0
Original source code for Miner 2149. This Source code is made available for your personal information and use only.  You may not modify or copy or distribute this code without the author's written permission. You may create a derivative work for private personal use only, not for distribution, promotion, or sale. Please contact the author if you are interested licensing for the creation versions of the program for modern devices, media, or other purposes.